Parameters: split varianter and parameters modules
Our goal to completely separate the basic parameter interface and the
varianter pass through here: splitting these modules into their
reespective ones.
