+++ avocado/core/varianter.py
@@ -16,301 +16,16 @@
 #          Lucas Meneghel Rodrigues <lmr@redhat.com>
 
 """
-Multiplex and create variants.
+Base classes for implementing the varianter interface
 """
 
 import hashlib
-import re
 
-from six import iterkeys, iteritems, itervalues
-from six.moves import xrange as range
+from six import iteritems, itervalues
 
 from . import tree
 from . import dispatcher
 from . import output
-
-
-class NoMatchError(KeyError):
-    pass
-
-
-class AvocadoParams(object):
-
-    """
-    Params object used to retrieve params from given path. It supports
-    absolute and relative paths. For relative paths one can define multiple
-    paths to search for the value.
-    It contains compatibility wrapper to act as the original avocado Params,
-    but by special usage you can utilize the new API. See ``get()``
-    docstring for details.
-
-    You can also iterate through all keys, but this can generate quite a lot
-    of duplicate entries inherited from ancestor nodes.  It shouldn't produce
-    false values, though.
-
-    In this version each new "get()" call is logged into avocado.LOG_JOB.
-    This is subject of change (separate file, perhaps)
-    """
-
-    # TODO: Use "test" to log params.get()
-
-    def __init__(self, leaves, test_id, mux_path, default_params):
-        """
-        :param leaves: List of TreeNode leaves defining current variant
-        :param test_id: test id
-        :param mux_path: list of entry points
-        :param default_params: dict of params used when no matches found
-
-        .. note:: `default_params` will be deprecated by the end of 2017.
-        """
-        self._rel_paths = []
-        leaves = list(leaves)
-        for i, path in enumerate(mux_path):
-            path_leaves = self._get_matching_leaves(path, leaves)
-            self._rel_paths.append(AvocadoParam(path_leaves,
-                                                '%d: %s' % (i, path)))
-        # Don't use non-mux-path params for relative paths
-        path_leaves = self._get_matching_leaves('/*', leaves)
-        self._abs_path = AvocadoParam(path_leaves, '*: *')
-        self.id = test_id
-        self._cache = {}     # TODO: Implement something more efficient
-        # TODO: Get rid of this and prepare something better
-        self._default_params = default_params
-
-    def __eq__(self, other):
-        if set(iterkeys(self.__dict__)) != set(iterkeys(other.__dict__)):
-            return False
-        for attr in iterkeys(self.__dict__):
-            if (getattr(self, attr) != getattr(other, attr)):
-                return False
-        return True
-
-    def __ne__(self, other):
-        return not (self == other)
-
-    def __getstate__(self):
-        """ log can't be pickled """
-        copy = self.__dict__.copy()
-        return copy
-
-    def __setstate__(self, orig):
-        """ refresh log """
-        self.__dict__.update(orig)
-
-    def __repr__(self):
-        return "<AvocadoParams %s>" % self._str()
-
-    def __str__(self):
-        return "params {%s}" % self._str()
-
-    def _str(self):
-        out = ",".join(_.str_leaves_variant for _ in self._rel_paths)
-        if out:
-            return self._abs_path.str_leaves_variant + ',' + out
-        else:
-            return self._abs_path.str_leaves_variant
-
-    def log(self, key, path, default, value):
-        """ Predefined format for displaying params query """
-        output.LOG_JOB.debug("PARAMS (key=%s, path=%s, default=%s) => %r", key,
-                             path, default, value)
-
-    def _get_matching_leaves(self, path, leaves):
-        """
-        Pops and returns list of matching nodes
-        :param path: Path (str)
-        :param leaves: list of TreeNode leaves
-        """
-        path = self._greedy_path(path)
-        path_leaves = [leaf for leaf in leaves if path.search(leaf.path + '/')]
-        for leaf in path_leaves:
-            leaves.remove(leaf)
-        return path_leaves
-
-    @staticmethod
-    def _greedy_path(path):
-        """
-        converts user-friendly asterisk path to python regexp and compiles it:
-        path = ""             => ^$
-        path = "/"            => /
-        path = "/foo/bar"     => /foo/bar
-        path = "foo/bar"      => $MUX_ENTRY/?.*/foo/bar
-        path = "/*/foo"       => /[^/]*/foo
-        path = "foo/*"        => $MUX_ENTRY/?.*/foo/.*
-        path = "/foo/*"       => /foo/.*
-        """
-        if not path:
-            return re.compile('^$')
-        if path[-1] == '*':
-            suffix = ''
-            path = path[:-1]
-        else:
-            suffix = '$'
-        return re.compile(path.replace('*', '[^/]*') + suffix)
-
-    @staticmethod
-    def _is_abspath(path):
-        """ Is this an absolute or relative path? """
-        if path.pattern and path.pattern[0] == '/':
-            return True
-        else:
-            return False
-
-    def __getattr__(self, attr):
-        """
-        Compatibility to old Params
-        :warning: This will be removed soon. Use params.get() instead
-        """
-        if attr == '__getnewargs__':    # pickling uses this attr
-            raise AttributeError
-        elif attr in self.__dict__:
-            return self.__dict__[attr]
-        else:
-            msg = ("You're probably retrieving param %s via attributes "
-                   " (self.params.$key) which is obsoleted. Use "
-                   "self.params.get($key) instead." % attr)
-            output.LOG_JOB.warn(msg)
-            return self.get(attr)
-
-    def get(self, key, path=None, default=None):
-        """
-        Retrieve value associated with key from params
-        :param key: Key you're looking for
-        :param path: namespace ['*']
-        :param default: default value when not found
-        :raise KeyError: In case of multiple different values (params clash)
-        """
-        if path is None:    # default path is any relative path
-            path = '*'
-        try:
-            return self._cache[(key, path, default)]
-        except (KeyError, TypeError):
-            # KeyError - first query
-            # TypeError - unable to hash
-            value = self._get(key, path, default)
-            self.log(key, path, default, value)
-            try:
-                self._cache[(key, path, default)] = value
-            except TypeError:
-                pass
-            return value
-
-    def _get(self, key, path, default):
-        """
-        Actual params retrieval
-        :param key: key you're looking for
-        :param path: namespace
-        :param default: default value when not found
-        :raise KeyError: In case of multiple different values (params clash)
-        """
-        path = self._greedy_path(path)
-        for param in self._rel_paths:
-            try:
-                return param.get_or_die(path, key)
-            except NoMatchError:
-                pass
-        if self._is_abspath(path):
-            try:
-                return self._abs_path.get_or_die(path, key)
-            except NoMatchError:
-                pass
-        return self._default_params.get(key, default)
-
-    def objects(self, key, path=None):
-        """
-        Return the names of objects defined using a given key.
-
-        :param key: The name of the key whose value lists the objects
-                (e.g. 'nics').
-        """
-        return self.get(path, key, "").split()
-
-    def iteritems(self):
-        """
-        Iterate through all available params and yield origin, key and value
-        of each unique value.
-        """
-        env = []
-        for param in self._rel_paths:
-            for path, key, value in param.iteritems():
-                if (path, key) not in env:
-                    env.append((path, key))
-                    yield (path, key, value)
-        for path, key, value in self._abs_path.iteritems():
-            if (path, key) not in env:
-                env.append((path, key))
-                yield (path, key, value)
-
-
-class AvocadoParam(object):
-
-    """
-    This is a single slice params. It can contain multiple leaves and tries to
-    find matching results.
-    """
-
-    def __init__(self, leaves, name):
-        """
-        :param leaves: this slice's leaves
-        :param name: this slice's name (identifier used in exceptions)
-        """
-        # Basic initialization
-        self._leaves = leaves
-        # names cache (leaf.path is quite expensive)
-        self._leaf_names = [leaf.path + '/' for leaf in leaves]
-        self.name = name
-
-    def __eq__(self, other):
-        if self.__dict__ == other.__dict__:
-            return True
-        else:
-            return False
-
-    def __ne__(self, other):
-        return not (self == other)
-
-    @property
-    def str_leaves_variant(self):
-        """ String with identifier and all params """
-        return "%s (%s)" % (self.name, self._leaf_names)
-
-    def _get_leaves(self, path):
-        """
-        Get all leaves matching the path
-        """
-        return [self._leaves[i]
-                for i in range(len(self._leaf_names))
-                if path.search(self._leaf_names[i])]
-
-    def get_or_die(self, path, key):
-        """
-        Get a value or raise exception if not present
-        :raise NoMatchError: When no matches
-        :raise KeyError: When value is not certain (multiple matches)
-        """
-        leaves = self._get_leaves(path)
-        ret = [(leaf.environment[key], leaf.environment.origin[key])
-               for leaf in leaves
-               if key in leaf.environment]
-        if not ret:
-            raise NoMatchError("No matches to %s => %s in %s"
-                               % (path.pattern, key, self.str_leaves_variant))
-        if len(set([_[1] for _ in ret])) == 1:  # single source of results
-            return ret[0][0]
-        else:
-            raise ValueError("Multiple %s leaves contain the key '%s'; %s"
-                             % (path.pattern, key,
-                                ["%s=>%s" % (_[1].path, _[0])
-                                 for _ in ret]))
-
-    def iteritems(self):
-        """
-        Very basic implementation which iterates through __ALL__ params,
-        which generates lots of duplicate entries due to inherited values.
-        """
-        for leaf in self._leaves:
-            for key, value in iteritems(leaf.environment):
-                yield (leaf.environment.origin[key].path, key, value)
 
 
 def is_empty_variant(variant):
